Improve scripts.

+++ setup/docker/python/satellite.sh
@@ -2,27 +2,32 @@ green=$'\e[0;32m'
 orange=$'\e[0;33m'
 white=$'\e[0m'
 
-echo "${green}Welcome to the satellite.sh file!${white}"
+echo "${green}Welcome to the satellite.sh script!${white}"
 
-# prepare python environment
+# virtual environment
+env="~/venv"
+
+if [ ! -d $env ]; then 
+    echo "Create virtual environment ${green}${env}${white}"
+    python3 -m venv $env 
+fi
+
+echo "Activate virtual environment ${green}${env}${white}"
+source $env/bin/activate
+
 if [ -f "requirements.txt" ]; then
-    echo "${green}Set up virtual environment${white}"
-    env="~/venv"
 
-    if [ ! -d $env ]; then 
-        python3 -m venv $env 
+    if [ -f "../commit_changed" ]; then
+        echo "${green}Install requirements${white}"
+        python -m pip install --pre --index-url https://www.myget.org/F/apollo3zehn-dev/python/ -r "requirements.txt" --disable-pip-version-check
     fi
 
-    source $env/bin/activate
-
-    echo "${green}Install requirements${white}"
-    python -m pip install --pre --index-url https://www.myget.org/F/apollo3zehn-dev/python/ -r "requirements.txt" --disable-pip-version-check
 else
     echo "${orange}No requirements.txt found${white}"
 fi
 
+# run user code
 echo "${green}Run user code${white}"
 
-# run user code (finally!)
 shift
-$@
+python $@
